Kontrolna tacka1 nisa

+++ InitialProject/InitialProject/View/Owner/MoveReservationView.xaml.cs
@@ -44,22 +44,30 @@ namespace TravelAgency.View.Owner
 
         private void AcceptMoveReservation(object sender, RoutedEventArgs e)
         {
-            SelectedReservation = (MoveReservation)DataPanel.SelectedItem;
-            reservationService.MoveReservation(SelectedReservation.ReservationId,SelectedReservation.NewStartDate,SelectedReservation.NewEndDate);
-            DataPanel.ItemsSource = moveReservationRepository.GetAll();
+            if (this.SelectedReservation != null) 
+            {
+                SelectedReservation = (MoveReservation)DataPanel.SelectedItem;
+                reservationService.MoveReservation(SelectedReservation.ReservationId, SelectedReservation.NewStartDate, SelectedReservation.NewEndDate);
+                DataPanel.ItemsSource = moveReservationRepository.GetAll();
+            }
+            else { }       
         }
 
         private void DeclineMoveReservation(object sender, RoutedEventArgs e)
         {
-            SelectedReservation = (MoveReservation)DataPanel.SelectedItem;
-            moveReservationRepository.Delete(SelectedReservation);
-            DataPanel.ItemsSource = moveReservationRepository.GetAll();
+            if(SelectedReservation != null) 
+            {
+                SelectedReservation = (MoveReservation)DataPanel.SelectedItem;
+                moveReservationRepository.Delete(SelectedReservation);
+                DataPanel.ItemsSource = moveReservationRepository.GetAll();
+            }
+            else { }   
         }
 
         private void ReservationInfo(object sender, SelectionChangedEventArgs e)
         {
             SelectedReservation = (MoveReservation)DataPanel.SelectedItem;
-            ReservationInfoLabel.Content = reservationService.TextForReservationInfo(SelectedReservation.HotelName,SelectedReservation.NewStartDate,SelectedReservation.NewEndDate);
+            ReservationInfoLabel.Content = reservationService.TextForReservationInfo(SelectedReservation.ReservationId, SelectedReservation.HotelName,SelectedReservation.NewStartDate,SelectedReservation.NewEndDate);
         }
     }
 }
